Destroy 컨트롤러 수정

+++ controllers/BookController.js
@@ -8,14 +8,14 @@ const getAllBookInformation = async (req, res) => {
         */
     });
 
-    return res.json({ success: true, data });
+    return res.json({ status: "success", data });
 }
 
 const createBookInformation = async (req, res) =>{
     req.body.remain_amount = req.body.amount;
     await Book.create(req.body);
 
-    return res.json({ success: true });
+    return res.json({ status: "success" });
 }
 
 const updateBookInformation = async (req, res) => {
@@ -32,15 +32,26 @@ const updateBookInformation = async (req, res) => {
         where: { book_id: req.params.bid }
     });
 
-    return res.json({ success: true });
+    return res.json({ status: "success" });
 }
 
 const deleteBookInformation = async (req, res) => {
-    await Book.destroy({
-        where: { book_id: req.params.bid }
+    const data = await Book.findOne({
+        where: { book_id: req.params.bid },
+        raw: true
     });
 
-    return res.json({ success: true });
+    if (data === null) {
+        return res.status(400).json({
+            status: "error",
+            message: "No content found for "
+        })
+    } else {
+        await Book.destroy({
+            where: { book_id: req.params.bid }
+        });
+        return res.json({ status: "success" });
+    }
 }
 
 module.exports = {
